Pr comment fix; Move graph abort transport out of the stable Metalium API. GraphFunctionAbort and ScopedTrackedFunction are capture-runtime plumbing, not a public processor contract; keep abort on track_function_end(std::any) and house the types under tt::tt_metal::internal.
