+++ tt_metal/impl/graph/graph_tracking.cpp
@@ -3,8 +3,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include <graph_tracking.hpp>
+#include <internal/graph_tracking.hpp>
 
 #include <algorithm>
+#include <any>
+#include <string>
+#include <string_view>
 #include <nlohmann/json.hpp>
 #include <tt_stl/assert.hpp>
 
@@ -31,24 +35,6 @@ bool GraphTracker::is_enabled() const {
 }
 
 bool GraphTracker::has_processors() const { return !processors.empty(); }
-
-void GraphTracker::track_function_abort(std::string_view reason) {
-    if (processors.empty()) {
-        return;
-    }
-    std::any payload{GraphFunctionAbort{std::string(reason), false}};
-    std::for_each(
-        processors.begin(), processors.end(), [&](auto& processor) { processor->track_function_end(payload); });
-}
-
-void GraphTracker::unwind_open_functions(std::string_view reason) {
-    if (processors.empty()) {
-        return;
-    }
-    std::any payload{GraphFunctionAbort{std::string(reason), true}};
-    std::for_each(
-        processors.begin(), processors.end(), [&](auto& processor) { processor->track_function_end(payload); });
-}
 
 void GraphTracker::push_processor(const std::shared_ptr<IGraphProcessor>& new_processor) {
     processors.push_back(new_processor);
@@ -256,3 +242,17 @@ void GraphTracker::clear_hook() {
 }
 
 }  // namespace tt::tt_metal
+
+namespace tt::tt_metal::internal {
+
+void unwind_open_functions(std::string_view reason) {
+    const auto& processors = GraphTracker::instance().get_processors();
+    if (processors.empty()) {
+        return;
+    }
+    std::any payload{GraphFunctionAbort{std::string(reason), true}};
+    std::for_each(
+        processors.begin(), processors.end(), [&](auto& processor) { processor->track_function_end(payload); });
+}
+
+}  // namespace tt::tt_metal::internal
